* Fixed cli_encryption_on_password statistic on 64bit

+++ libsrc/Wi/security.c
@@ -2129,7 +2129,7 @@ sec_stmt_exec (query_instance_t * qi, ST * tree)
     }
 }
 
-int32 cli_encryption_on_password = 0; /* digest */
+long cli_encryption_on_password = 0; /* digest */
 static caddr_t caller_id_defaults = NULL;
 
 static void
